+++ .eslintrc
@@ -23,7 +23,6 @@
     "react/jsx-uses-vars": "error",
     "react/react-in-jsx-scope": "error",
     "no-underscore-dangle": "off",
-    "import/no-extraneous-dependencies": "off",
-    # "linebreak-style": "off"
+    "import/no-extraneous-dependencies": "off"
   }
 }
