+++ app/containers/app.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import AddNote from './note/add-note.jsx';
-import NotesContainer from './note/notes';
+import NotesContainer from './note/notes.jsx';
 
 const App = () => (
   <section className="container">
@@ -8,4 +8,5 @@ const App = () => (
     <AddNote/>
   </section>
 );
+
 export default App;
